% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.r
\name{repr-generics}
\alias{repr-generics}
\alias{repr_html}
\alias{repr_html.default}
\alias{repr_markdown}
\alias{repr_markdown.default}
\alias{repr_latex}
\alias{repr_latex.default}
\alias{repr_json}
\alias{repr_json.default}
\alias{repr_javascript}
\alias{repr_javascript.default}
\alias{repr_pdf}
\alias{repr_pdf.default}
\alias{repr_png}
\alias{repr_png.default}
\alias{repr_jpg}
\alias{repr_jpg.default}
\alias{repr_svg}
\alias{repr_svg.default}
\title{Representations for specific formats}
\usage{
repr_html(obj, ...)

\method{repr_html}{default}(obj, ...)

repr_markdown(obj, ...)

\method{repr_markdown}{default}(obj, ...)

repr_latex(obj, ...)

\method{repr_latex}{default}(obj, ...)

repr_json(obj, ...)

\method{repr_json}{default}(obj, ...)

repr_javascript(obj, ...)

\method{repr_javascript}{default}(obj, ...)

repr_pdf(obj, ...)

\method{repr_pdf}{default}(obj, ...)

repr_png(obj, ...)

\method{repr_png}{default}(obj, ...)

repr_jpg(obj, ...)

\method{repr_jpg}{default}(obj, ...)

repr_svg(obj, ...)

\method{repr_svg}{default}(obj, ...)
}
\arguments{
\item{obj}{The object to create a repr for}

\item{...}{parameters of the specific \code{repr_*} functions}
}
\description{
Representations for specific formats
}
\seealso{
\link{repr_text} for the only repr that is always defined
}
