% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.r
\docType{package}
\name{repr-package}
\alias{repr-package}
\title{The repr package}
\description{
The repr package
}
\details{
The LaTeX repr of functions with the \code{repr.function.highlight} option set to \code{FALSE} and the \code{repr.function.highlightlatex} option set to \code{TRUE} needs \code{\\usepackage{minted}}
}
\seealso{
\link{repr}, \link{repr-options}, \link{repr-generics}, \link{repr_text}
}

