% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.r
\name{repr}
\alias{repr}
\title{Dynamic representation}
\usage{
repr(obj, format = "text", ...)
}
\arguments{
\item{obj}{The object to create a representation for}

\item{format}{The representation format. \code{repr_<format>} is then called. (default: Call \link{repr_text})}

\item{...}{delegated to the specific \code{repr_<format>} function}
}
\value{
A character or raw vector of that format or NULL if none is defined.
Only the \code{'text'} format is defined for everything (via \link{print}\code{()})
}
\description{
Specify an object and a format to represent it in. Will \link{stop}\code{()} if no such format is known.
}
\seealso{
\link{repr_text}, \link{repr-generics}
}

