% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repr_datatable.r
\name{repr_*.data.table}
\alias{repr_*.data.table}
\alias{repr_html.data.table}
\alias{repr_text.data.table}
\alias{repr_latex.data.table}
\title{Representation of data.table objects}
\usage{
\method{repr_html}{data.table}(obj, ...)

\method{repr_text}{data.table}(obj, ...)

\method{repr_latex}{data.table}(obj, ...)
}
\arguments{
\item{obj}{The list to create a representation for}

\item{...}{ignored}
}
\description{
Representation of data.table objects
}
