% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repr_factor.r
\name{repr_*.factor}
\alias{repr_*.factor}
\alias{repr_html.factor}
\alias{repr_markdown.factor}
\alias{repr_latex.factor}
\title{Representations of factors}
\usage{
\method{repr_html}{factor}(obj, ...)

\method{repr_markdown}{factor}(obj, ...)

\method{repr_latex}{factor}(obj, ...)
}
\arguments{
\item{obj}{The factor to create a representation for}

\item{...}{ignored}
}
\description{
Representations of factors
}
