% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repr_function.r
\name{repr_*.function}
\alias{repr_*.function}
\alias{repr_html.function}
\alias{repr_latex.function}
\alias{repr_markdown.function}
\title{Representations of functions}
\usage{
\method{repr_html}{function}(obj,
  highlight = getOption("repr.function.highlight"), ...)

\method{repr_latex}{function}(obj,
  highlight = getOption("repr.function.highlight"), ...)

\method{repr_markdown}{function}(obj, fenced = TRUE, ...)
}
\arguments{
\item{obj}{Function to create a representation for}

\item{highlight}{Should code highlighting be performed}

\item{...}{ignored}

\item{fenced}{Should a fenced code block instead of an indented one be used?}
}
\description{
Representations of functions
}
