% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repr_help_files_with_topic.r
\name{repr_*.help_files_with_topic}
\alias{repr_*.help_files_with_topic}
\alias{repr_text.help_files_with_topic}
\alias{repr_html.help_files_with_topic}
\alias{repr_latex.help_files_with_topic}
\title{Representations of help}
\usage{
\method{repr_text}{help_files_with_topic}(obj, ...)

\method{repr_html}{help_files_with_topic}(obj, ...)

\method{repr_latex}{help_files_with_topic}(obj, ...)
}
\arguments{
\item{obj}{Help topic to create a representation for}

\item{...}{ignored}
}
\description{
Representations of help
}
