% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repr_htmlwidget.r
\docType{data}
\name{repr_*.htmlwidget}
\alias{repr_*.htmlwidget}
\alias{html_dependencies}
\alias{repr_text.htmlwidget}
\alias{repr_html.htmlwidget}
\alias{repr_text.shiny.tag}
\alias{repr_*.shiny.tag}
\alias{repr_html.shiny.tag}
\alias{repr_text.shiny.tag.list}
\alias{repr_*.shiny.tag.list}
\alias{repr_html.shiny.tag.list}
\title{HTML widget representations}
\format{
An object of class \code{environment} of length 4.
}
\usage{
html_dependencies

\method{repr_text}{htmlwidget}(obj, ...)

\method{repr_html}{htmlwidget}(obj, ...)

\method{repr_text}{shiny.tag}(obj, ...)

\method{repr_html}{shiny.tag}(obj, ...)

\method{repr_text}{shiny.tag.list}(obj, ...)

\method{repr_html}{shiny.tag.list}(obj, ...)
}
\arguments{
\item{obj}{The htmlwidget, shiny.tag, or shiny.tag.list to create a representation for}

\item{...}{ignored}
}
\description{
Standalone HTML representation and dummy text representation.
}
\details{
\code{html_dependencies} is an \link[base]{environment} containing the following functions.
\code{getOption(\link[=repr-options]{'repr.html.deduplicate'})}
\describe{
 \item{\code{get()}}{Get the list of added dependencies}
 \item{\code{add(dep)}}{Marks a dependency as added. Call this e.g. after appending a script tag with the dependency.}
 \item{\code{clear()}}{Clear the list as seen dependencies. Now everything will be added again when encountered.}
 \item{\code{dir()}}{Returns the directory in which the dependencies reside.}
}
}
\keyword{datasets}
