% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repr_list.r
\name{repr_*.list}
\alias{repr_*.list}
\alias{repr_html.list}
\alias{repr_markdown.list}
\alias{repr_latex.list}
\title{Representations of lists}
\usage{
\method{repr_html}{list}(obj, ...)

\method{repr_markdown}{list}(obj, ...)

\method{repr_latex}{list}(obj, ...)
}
\arguments{
\item{obj}{The list to create a representation for}

\item{...}{ignored}
}
\description{
Representations of lists
}
