% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repr_packageIQR.r
\name{repr_*.packageIQR}
\alias{repr_*.packageIQR}
\alias{repr_html.packageIQR}
\alias{repr_text.packageIQR}
\title{packageIQR representations}
\usage{
\method{repr_text}{packageIQR}(obj, ...)

\method{repr_html}{packageIQR}(obj, ...)
}
\arguments{
\item{obj}{The packageIQR obj to create a representation for}

\item{...}{ignored}
}
\description{
plain text representations of packageIQR objects like the list of vailable example 
data or vignettes
}

