% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repr_htmlwidget.r
\name{repr_*.shiny.tag}
\alias{repr_*.shiny.tag}
\alias{repr_text.shiny.tag}
\alias{repr_html.shiny.tag}
\title{Shiny tag representations}
\usage{
\method{repr_text}{shiny.tag}(obj, ...)

\method{repr_html}{shiny.tag}(obj, ...)
}
\arguments{
\item{obj}{The shiny tags to create a representation for}

\item{...}{ignored}
}
\description{
Standalone HTML representation and dummy text representation
}
