% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repr_ts.r
\name{repr_*.ts}
\alias{repr_*.ts}
\alias{repr_html.ts}
\alias{repr_latex.ts}
\alias{repr_markdown.ts}
\alias{repr_text.ts}
\title{Time series representations}
\usage{
\method{repr_html}{ts}(obj, ...)

\method{repr_latex}{ts}(obj, ...,
  colspec = getOption("repr.matrix.latex.colspec"))

\method{repr_markdown}{ts}(obj, ...)

\method{repr_text}{ts}(obj, ...)
}
\arguments{
\item{obj}{The \link[stats]{ts} object to create a representation for}

\item{...}{ignored}

\item{colspec}{The colspec for the LaTeX table. The default is given by the option \code{repr.matrix.latex.colspec}}
}
\description{
HTML, LaTeX, and Markdown representations of \link[stats]{ts} objects.
}
\seealso{
\link{repr-options} for \code{repr.matrix.latex.colspec}
}
