% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repr_spatial.r
\name{repr_geojson.*}
\alias{repr_geojson.*}
\alias{repr_geojson.geo_list}
\alias{repr_geojson.SpatialCollections}
\alias{repr_geojson.SpatialPolygons}
\alias{repr_geojson.SpatialPolygonsDataFrame}
\alias{repr_geojson.SpatialPoints}
\alias{repr_geojson.SpatialPointsDataFrame}
\alias{repr_geojson.SpatialLines}
\alias{repr_geojson.SpatialLinesDataFrame}
\alias{repr_geojson.SpatialGrid}
\alias{repr_geojson.SpatialGridDataFrame}
\alias{repr_geojson.SpatialPixels}
\alias{repr_geojson.SpatialPixelsDataFrame}
\alias{repr_geojson.SpatialRings}
\alias{repr_geojson.SpatialRingsDataFrame}
\alias{repr_geojson.sf}
\alias{repr_geojson.sfg}
\alias{repr_geojson.sfc}
\title{Representations of spatial objects: See \link[geojsonio]{geojson_list} for supported classes.}
\usage{
\method{repr_geojson}{geo_list}(obj, ...)

\method{repr_geojson}{SpatialCollections}(obj, ...)

\method{repr_geojson}{SpatialPolygons}(obj, ...)

\method{repr_geojson}{SpatialPolygons}(obj, ...)

\method{repr_geojson}{SpatialPolygonsDataFrame}(obj, ...)

\method{repr_geojson}{SpatialPoints}(obj, ...)

\method{repr_geojson}{SpatialPointsDataFrame}(obj, ...)

\method{repr_geojson}{SpatialLines}(obj, ...)

\method{repr_geojson}{SpatialLinesDataFrame}(obj, ...)

\method{repr_geojson}{SpatialGrid}(obj, ...)

\method{repr_geojson}{SpatialGridDataFrame}(obj, ...)

\method{repr_geojson}{SpatialPixels}(obj, ...)

\method{repr_geojson}{SpatialPixelsDataFrame}(obj, ...)

\method{repr_geojson}{SpatialRings}(obj, ...)

\method{repr_geojson}{SpatialRingsDataFrame}(obj, ...)

\method{repr_geojson}{sf}(obj, ...)

\method{repr_geojson}{sfg}(obj, ...)

\method{repr_geojson}{sfc}(obj, ...)
}
\arguments{
\item{obj}{The spatial object to create a representation for}

\item{...}{ignored}
}
\description{
Representations of spatial objects: See \link[geojsonio]{geojson_list} for supported classes.
}
