% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repr_plotly.r
\name{repr_plotly1.*}
\alias{repr_plotly1.*}
\alias{repr_plotly1.plotly}
\alias{repr_plotly1.gg}
\title{Representation as \link[plotly:plotly_json]{Plotly JSON}.}
\usage{
\method{repr_plotly1}{plotly}(obj, ...)

\method{repr_plotly1}{gg}(obj, ...)
}
\arguments{
\item{obj}{The \link[plotly]{plot_ly} plot or \link[ggplot2]{ggplot} to create a representation for}

\item{...}{ignored}
}
\description{
Representation as \link[plotly:plotly_json]{Plotly JSON}.
}
