% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.r
\name{repr_text}
\alias{repr_text}
\alias{repr_text.default}
\title{Text representation}
\usage{
repr_text(obj, ...)

\method{repr_text}{default}(obj, ...)
}
\arguments{
\item{obj}{The object to \link{print} and then return the output}

\item{...}{ignored}
}
\description{
The only representation defined per default for everthing (via \link{print}\code{()})
}
\seealso{
\link{repr-generics} for other generics
}

