% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repr_vega.r
\name{repr_vega(4/lite2).*}
\alias{repr_vega(4/lite2).*}
\alias{repr_vegalite2.vegalite}
\title{Representation as \code{\link[vegalite]{vegalite}v2} or \code{vega4} JSON.}
\usage{
\method{repr_vegalite2}{vegalite}(obj, ...)
}
\arguments{
\item{obj}{The \link[vegalite]{vegalite} plot to create a representation for}

\item{...}{ignored}
}
\description{
Representation as \code{\link[vegalite]{vegalite}v2} or \code{vega4} JSON.
}
