% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.r
\docType{package}
\name{repr-package}
\alias{repr-package}
\title{The repr package}
\description{
String and binary representations of objects for several formats / mime types.
}
\details{
The LaTeX repr of vectors needs \code{\\usepackage[inline]{enumitem}}

The LaTeX repr of functions with the \code{repr.function.highlight} option set to FALSE needs \code{\\usepackage{minted}}
}
\seealso{
\link{repr}, \link{repr-options}, \link{repr-generics}, \link{repr_text}
}
