% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repr_htmlwidget.r
\name{repr_*.htmlwidget}
\alias{repr_*.htmlwidget}
\alias{repr_text.htmlwidget}
\alias{repr_html.htmlwidget}
\alias{repr_text.shiny.tag.list}
\alias{repr_html.shiny.tag.list}
\title{HTML widget representations}
\usage{
\method{repr_text}{htmlwidget}(obj, ...)

\method{repr_html}{htmlwidget}(obj, ...)

\method{repr_text}{shiny.tag.list}(obj, ...)

\method{repr_html}{shiny.tag.list}(obj, ...)
}
\arguments{
\item{obj}{The htmlwidget to create a representation for}

\item{...}{ignored}
}
\description{
Standalone HTML representation and dummy text representation
}
