% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repr_htmlwidget.r
\name{repr_*.shiny.tag.list}
\alias{repr_*.shiny.tag.list}
\alias{repr_text.shiny.tag.list}
\title{Standalone HTML representation and dummy text representation}
\usage{
\method{repr_text}{shiny.tag.list}(obj, ...)
}
\arguments{
\item{obj}{The shiny tags to create a representation for}

\item{...}{ignored}
}
\description{
Standalone HTML representation and dummy text representation
}
