% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repr_vector.r
\name{repr_*.vector}
\alias{repr_*.vector}
\alias{repr_html.logical}
\alias{repr_html.integer}
\alias{repr_html.complex}
\alias{repr_html.numeric}
\alias{repr_html.character}
\alias{repr_html.Date}
\alias{repr_markdown.logical}
\alias{repr_markdown.integer}
\alias{repr_markdown.complex}
\alias{repr_markdown.numeric}
\alias{repr_markdown.character}
\alias{repr_markdown.Date}
\alias{repr_latex.logical}
\alias{repr_latex.integer}
\alias{repr_latex.complex}
\alias{repr_latex.numeric}
\alias{repr_latex.character}
\alias{repr_latex.Date}
\title{Representations of vectors}
\usage{
\method{repr_html}{logical}(obj, ..., items = getOption("repr.vector.max.items"))

\method{repr_html}{integer}(obj, ..., items = getOption("repr.vector.max.items"))

\method{repr_html}{complex}(obj, ..., items = getOption("repr.vector.max.items"))

\method{repr_html}{numeric}(obj, ..., items = getOption("repr.vector.max.items"))

\method{repr_html}{character}(obj, ..., items = getOption("repr.vector.max.items"))

\method{repr_html}{Date}(obj, ..., items = getOption("repr.vector.max.items"))

\method{repr_markdown}{logical}(obj, ..., items = getOption("repr.vector.max.items"))

\method{repr_markdown}{integer}(obj, ..., items = getOption("repr.vector.max.items"))

\method{repr_markdown}{complex}(obj, ..., items = getOption("repr.vector.max.items"))

\method{repr_markdown}{numeric}(obj, ..., items = getOption("repr.vector.max.items"))

\method{repr_markdown}{character}(obj, ..., items = getOption("repr.vector.max.items"))

\method{repr_markdown}{Date}(obj, ..., items = getOption("repr.vector.max.items"))

\method{repr_latex}{logical}(obj, ..., items = getOption("repr.vector.max.items"))

\method{repr_latex}{integer}(obj, ..., items = getOption("repr.vector.max.items"))

\method{repr_latex}{complex}(obj, ..., items = getOption("repr.vector.max.items"))

\method{repr_latex}{numeric}(obj, ..., items = getOption("repr.vector.max.items"))

\method{repr_latex}{character}(obj, ..., items = getOption("repr.vector.max.items"))

\method{repr_latex}{Date}(obj, ..., items = getOption("repr.vector.max.items"))
}
\arguments{
\item{obj}{The vector to create a representation for}

\item{...}{ignored}

\item{items}{The maximum number of items displayed. The default is given by the option \code{repr.vector.max.items}}
}
\description{
Representations of vectors
}
