% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.r
\docType{data}
\name{*2repr}
\alias{*2repr}
\alias{mime2repr}
\alias{format2repr}
\title{Lists mapping mime types (\code{mime2repr}) or format names (\code{format2repr}) to \code{repr} functions}
\format{Lists of length 10 mapping mime/name to function}
\usage{
mime2repr

format2repr
}
\description{
Lists mapping mime types (\code{mime2repr}) or format names (\code{format2repr}) to \code{repr} functions
}
\examples{
names(mime2repr)
names(format2repr)

}
\keyword{datasets}
